Diff.py: Convert incoming "file_list" on __init__
Tests of Diff functions are done with arrays, however the Diffs are
mostly initialized with tuples (files are read into tuples).
The changes done to "has_changes" before do not work with tuples,
which is not caught by the tests.
Converting the incoming tuple (or list) enables these changes to work
outside of tests.
Additionally, this change enables independence between the diff and
the array its constructed from. With the use of tuples this is already
given (tuple immutability) but with lists it isn't.